requirements.txt
Flask>=2.2.0,<4
Flask-Cors>=3.0.0,<5
python-dotenv>=0.21.0,<2
openai>=1.102.0
gunicorn>=23.0.0
PyMuPDF>=1.20.0,<2
Pillow>=9.0.0,<11
pytesseract>=0.3.10,<1
docx2txt>=0.8,<1
tiktoken>=0.4.0,<1
uvicorn>=0.35.0
google-cloud-aiplatform>=1.58.0
